Refactor fee management structure and related classes
Removed `FeeItem` and introduced `FeeProductItem` and `FeeReceiptItem` to better encapsulate fee receipt details and associated products. Updated `FeeModel` to `FeeProductModel` and `FeeReceiptModel` for modular management. Adjusted `ListFeeController`, `ListFeePage.xaml`, and `SaveFeeController` to work with the new structures. Enhanced `ResumeModel` and introduced `ReceiptCascadeOperationModel` and `ReceiptHandlerModel` for improved receipt handling. Overall architecture refactored for clarity and maintainability.

+++ src/receipts/ReceiptModel.cs
@@ -44,7 +44,7 @@ public class ReceiptModel : ABaseModel,
             if (reader.Read()) {
                 return new ReceiptItem {
                     receipt_id = reader.GetSafeValue<int>(m_COL_ID),
-                    receipt_total_price = reader.GetSafeValue<double>(m_COL_RECEIPT_TOTAL_PRICE),
+                    receipt_total_price = reader.GetSafeValue<decimal>(m_COL_RECEIPT_TOTAL_PRICE),
                     receipt_date_created = reader.GetSafeValue(m_COL_RECEIPT_DATE_CREATED, DateOnly.MinValue).ToString()
                 };
             }
